Update Background_Parallax.cs
1) upgraded the parallax script to dynamically determine the parallax factor by using the z axis position of the given layer.
1.1) depreciated brute force Method

note: it may boost performance to have each layer attached to it's horizontal copies and parented somehow. this way each chunk does not need to manually calculate it's position and can be moved by the parent object. this prevents the runtime from scaling up as the background gets wider. the runtime would only increase as the number of layers increase.