Upgraded Forest Background
1) implemented a new system for BG extension and parallax.
1.1) the new extension system is based on a single layer as opposed to the previous one which was based on multiple layered background chunks at once. the new system allows us to parent all of the copies of the base background layer to a single extender object. this ties into the parallax script to allowing us to shift every connected background layer segment at once rather calculating the movement for each background chunk individually.

this allows us to take prefabs for each chunk of a background layer and pass them to a Background extender that can patch them together as children of the extender.

We can then add each Extendable layer into a prefab for the Background as a whole, placing them in the appropriate order based on the objects desired z position. then you must manually set the z position of each extendable layer.

This change overall has increased the avg framerate form ~120fps to ~320fps.

The Runtime scaling is improved from O(n) where n is the number of chunks of background within the scene to O(1). (runtime is still impacted by adding additional layers to the background)